Clean up existing Indicator admin
for mercycorps/TolaActivity#2822

- collect all Admin models in admin.py
- organize for readability
